+++ crates/djls-project/src/lib.rs
@@ -1,3 +1,4 @@
+mod cache;
 mod db;
 mod discover;
 mod django;
@@ -8,6 +9,8 @@ mod python;
 mod resolve;
 mod symbols;
 
+pub use cache::load_cached_inspector_response;
+pub use cache::save_inspector_response;
 pub use db::Db;
 pub use discover::discover_template_libraries;
 pub use django::template_dirs;